refactor(ui): move key routing out of model.go
Dispatch by mode and the ordered dashboard handlers are the entry point for
every keystroke, so they read better in one place than interleaved with the
actions they call.

Declarations move unchanged.
